log all reflog per branch
The easiest way would be to find the head commit of the branch as it was
immediately before the rebase started in the reflog...

git reflog
and to reset the current branch to it (with the usual caveats about
being absolutely sure before reseting with the --hard option).

\# Suppose the old commit was HEAD@{5} in the ref log
git reset --hard HEAD@{5}
You can check the history of the candidate old head by just doing a git
log HEAD@{5} .

If you've enabled per branch reflogs you should be able to simply do git
reflog branchname@{1} as a rebase detaches the branch head before
reattaching to the final head.